+++ wheat-modern/src/main/scala/x7c1/wheat/modern/sequence/SequenceMapping.scala
@@ -13,3 +13,11 @@ trait SequenceMapping[A, F[_] <: Sequence[_]]{
 trait CanMapFrom[F[_]] {
   def mapFrom[A, B](fa: F[A])(f: A => B): F[B]
 }
+
+private[sequence] class DefaultCanMapFrom extends CanMapFrom[Sequence]{
+  override def mapFrom[A, B](fa: Sequence[A])(f: A => B) =
+    new Sequence[B] {
+      override def findAt(position: Int) = fa findAt position map f
+      override def length = fa.length
+    }
+}
